backend/requirements.txt
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.7
django-allauth==0.40.0
django-cors-headers==3.1.1
django-rest-auth==0.9.5
djangorestframework==3.10.3
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.7.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-core==1.1.0
google-cloud-translate==2.0.0
googleapis-common-protos==1.6.0
grpcio==1.25.0
httplib2==0.14.0
idna==2.8
oauthlib==3.1.0
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pymongo==3.9.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
